+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * Add type column to users settings permissions sets. Refs UIU-2167.
 * Add modal for 'Cash drawer reconciliation report'. Refs UIU-1959.
 * Provide `key` when rendering list elements. Refs UIU-2168.
+* Create 'Cash drawer reconciliation report' in PDF format. Refs UIU-2084.
 
 ## [6.0.0](https://github.com/folio-org/ui-users/tree/v6.0.0) (2021-03-18)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v5.0.9...v6.0.0)